requirements.txt
python-decouple==3.8

pymysql==1.1.1
cryptography==43.0.1
pycparser==2.22
asgiref==3.8.1
cffi==1.17.1
Django==5.1
djangorestframework==3.15.2
pip==24.0
PyJWT==2.9.0
sqlparse==0.5.1
gunicorn==20.1.0
